Documentation and window events
+ 'window' package
Contains more of the same callback/listener code as the input package.
This'll let us pause the game when the window's minimized and stuff like that.

+ Documentation update
I saw that you tooted my horn for me, so I went back through the classes I wrote and put my big, stupid name on top of all of them.
In addition, all of the Listeners and Callbacks have some decent JavaDoc comments now. We didn't really need them, but some of the event semantics are bizarre. WindowClose events should actually be called WindowClosing, for instance.

- KeyCallback.KEYBOARD
I knew that this was the solution to work with, ( since we need static access for your Component system ) but I wasn't sure where to shove it initially. I've since put it in Application, which seems closer to the actual data. If you have a better suggestion for where to plunk this guy, be my guest, provided it's not back in KeyboardCallback.

NOTES:
When I run the app, it hangs for a bit, then white-screens and sits there. I assume that's a symptom of twiddling with the rendering code? We'll get it straightened out.

- Erik