Wavefront OBJ Loading refactor, misc changes
+ OBJLoader
I don't know which implementation is better anymore, but I tested my OBJLoader on all our models and it worked.
I think I can tweak it to be even more useful, maybe with some pre-processing functions. ( I have to make one pass to begin with so that '/' characters are separated by whitespace before being parsed, so extending that step into a preprocessor wouldn't be too hard )

*ArrayList
I parameterized the growth factor and the initial capacity to more closely resemble ArrayList.java.
This way, we can create minimal array lists of fixed size without allocating wasteful space.

~  Lots of other files
I changed the name of a function in IntArrayList from asArray or something like it to 'toArray'. All the other changes should be related to that change or general formatting.

NOTE: Something isn't letting GLFW terminate again.
If you look at Application.java, there's a KeyboardListener bound to KEYBOARD that should quit the app when <ESC> is pressed, and the evaluation gets made, but the window doesn't close itself. To quit, you need to x out the window through the frame. I'll investigate further.

-Erik